Fixed a bug where unsuccessful ecrt_slave_config_reg_pdo_entry()'s went unnoticed and caused a SEGV

+++ rtw/lib/ecrt_support.h
@@ -15,7 +15,7 @@ struct pdo_map {
 
     /* The next values are used by the support layer */
     struct ecat_domain *domain;
-    size_t offset;
+    int offset;
     unsigned int bit_pos;
 };
